test cleanup: move helper .c and .h files to test/helpers/
