+++ test/helpers/cmp_testlib.c
