+++ test/recordlentest.c
@@ -9,7 +9,7 @@
 
 #include <string.h>
 
-#include "ssltestlib.h"
+#include "helpers/ssltestlib.h"
 #include "testutil.h"
 
 static char *cert = NULL;
