+++ test/ssl_test_ctx_test.c
@@ -16,7 +16,7 @@
 #include <string.h>
 
 #include "internal/nelem.h"
-#include "ssl_test_ctx.h"
+#include "helpers/ssl_test_ctx.h"
 #include "testutil.h"
 #include <openssl/e_os2.h>
 #include <openssl/err.h>
